refactor: split stockdetailmodal
