+++ src/main/resources/i18n/messages_en.properties
@@ -531,6 +531,7 @@ stockDetail.trend.up=Up
 stockDetail.trend.down=Down
 stockDetail.chartTitle=Price history
 stockDetail.weekRange= {0}\u2013{1}
+stockDetail.allTimeChange=All-time change
 stockDetail.historyTitle=Historical prices
 stockDetail.weeklyTitle=Weekly changes
 stockDetail.weeklyEmpty=No price changes yet
