+++ src/main/resources/i18n/messages_no.properties
@@ -527,6 +527,7 @@ stockDetail.weeklyEmpty=Ingen kursendringer ennå
 stockDetail.changeNative=Endring ({0})
 stockDetail.changeNok=Endring (NOK)
 stockDetail.changePct=Endring %
+stockDetail.allTimeChange=Alltime endring
 stockDetail.explore=Utforsk andre aksjer \u2192
 stockDetail.trend.flat=Uendret
 # Watchlist - dialog
